Bị lỗi lấy data cho today

+++ app/src/main/java/club/mobile/d21/personalassistant/ui/all_note/AllNoteFragment.kt
@@ -1,40 +1,50 @@
 package club.mobile.d21.personalassistant.ui.all_note
 
+import android.app.AlertDialog
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import androidx.fragment.app.activityViewModels
 import androidx.fragment.app.viewModels
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import club.mobile.d21.personalassistant.R
-import club.mobile.d21.personalassistant.data.Note
 import club.mobile.d21.personalassistant.databinding.FragmentAllNoteBinding
 import club.mobile.d21.personalassistant.ui.adapter.NoteAdapter
+import club.mobile.d21.personalassistant.ui.all_task.AllTaskViewModel
 import java.time.LocalDate
 
 class AllNoteFragment : Fragment() {
     private var _binding: FragmentAllNoteBinding? = null
     private val binding get() = _binding!!
-    private val allNoteViewModel: AllNoteViewModel by viewModels()
+    private val allNoteViewModel: AllNoteViewModel by activityViewModels()
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
         _binding = FragmentAllNoteBinding.inflate(inflater, container, false)
-        val notelist: RecyclerView = binding.noteList
+        val noteList: RecyclerView = binding.noteList
         allNoteViewModel.note.observe(viewLifecycleOwner) { list
-            -> notelist.adapter = NoteAdapter(list,
+            -> noteList.adapter = NoteAdapter(list,
             onEditClick = { selectedNote ->
                 val bottomSheet = EditNoteBottomSheet(selectedNote)
-                bottomSheet.show(childFragmentManager, bottomSheet.tag) },
-            onDeleteClick = {selectedNote->
-                allNoteViewModel.deleteNote(selectedNote)
-            })
+                bottomSheet.show(childFragmentManager, bottomSheet.tag) }
+            ) { selectedNote ->
+                val alertDialogBuilder = AlertDialog.Builder(requireContext())
+                alertDialogBuilder.setTitle("Confirm")
+                alertDialogBuilder.setMessage("Are you sure you want to delete this note?")
+                alertDialogBuilder.setPositiveButton("YES") { _, _ ->
+                    allNoteViewModel.deleteNote(selectedNote)
+                }
+                alertDialogBuilder.setNegativeButton("NO") { _, _ -> }
+                val alertDialog = alertDialogBuilder.create()
+                alertDialog.show()
+            }
         }
-        notelist.layoutManager = LinearLayoutManager(context)
+        noteList.layoutManager = LinearLayoutManager(context)
         return binding.root
     }
